Renaming extension and adding support for .rel files
