+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+### 1.0.7
+
+ - Renamed the langauge from Delve to Rel
+ - Files with .rel extension are also supported
+
 ### 1.0.6
 
  - Added support for `@auto_number`
